Removed bulk of ISharedService
Got the core window / tab management code moved into the `VsVimHost`
layer directly instead of through `ISharedService`

This took _much_ longer than it should have because I ended up
struggling with referencing the non-SDK DLLs. The
`Microsoft.VisualStudio.SDK` package by default will end up forcing many
of the PIA assemblies to be
`<EmbeddedInteropTypes>true</EmbeddedInteropTypes>`.

That works fine within the supported SDK but breaks down once you get to
DLLs like `Microsoft.VisualStudio.Platform.WindowManagement`. That,
appears at least, to uses a standard reference not an embedded one.
Hence it causes a conflict at compile time.

Took me a while to figure out how to suppress the interop type embedding
so that I could add the reference.
