+++ Src/VsSpecific/VsSpecific/SharedService.cs
@@ -34,50 +34,6 @@ namespace Vim.VisualStudio.Specific
             GetActiveViews()[index].ShowInFront();
         }
 
-        internal WindowFrameState GetWindowFrameState()
-        {
-            var activeView = ViewManager.Instance.ActiveView;
-            if (activeView == null)
-            {
-                return WindowFrameState.Default;
-            }
-
-            var list = GetActiveViews();
-            var index = list.IndexOf(activeView);
-            if (index < 0)
-            {
-                return WindowFrameState.Default;
-            }
-
-            return new WindowFrameState(index, list.Count);
-        }
-
-        /// <summary>
-        /// Get the list of View's in the current ViewManager DocumentGroup
-        /// </summary>
-        private static List<View> GetActiveViews()
-        {
-            var activeView = ViewManager.Instance.ActiveView;
-            if (activeView == null)
-            {
-                return new List<View>();
-            }
-
-            var group = activeView.Parent as DocumentGroup;
-            if (group == null)
-            {
-                return new List<View>();
-            }
-
-            return group.VisibleChildren.OfType<View>().ToList();
-        }
-
-        internal bool IsActiveWindowFrame(IVsWindowFrame vsWindowFrame)
-        {
-            var frame = vsWindowFrame as WindowFrame;
-            return frame != null && frame.FrameView == ViewManager.Instance.ActiveView;
-        }
-
         #region ISharedService
 
         WindowFrameState ISharedService.GetWindowFrameState()
